Address review round 2: ConditionRating float64, i18n aria-labels, label cleanup with IN query
Agent-Logs-Url: https://github.com/sj-tech-sweden/warehousecore/sessions/8773a636-f5d3-46e1-8cc3-f8a8309efa8b

+++ internal/handlers/device_admin_handlers.go
@@ -326,10 +326,10 @@ func BulkUpdateDevices(w http.ResponseWriter, r *http.Request) {
 	var req struct {
 		IDs     []string `json:"ids"`
 		Updates struct {
-			Status          *string `json:"status"`
-			ZoneID          *int    `json:"zone_id"`
-			CurrentLocation *string `json:"current_location"`
-			ConditionRating *int    `json:"condition_rating"`
+			Status          *string  `json:"status"`
+			ZoneID          *int     `json:"zone_id"`
+			CurrentLocation *string  `json:"current_location"`
+			ConditionRating *float64 `json:"condition_rating"`
 		} `json:"updates"`
 	}
 	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
